Conflicts:
	Compatibility.txt
	EGE/include/ege/base.h
	EGE/include/ege/ctl.h
	EGE/include/ege/gdi.h
	EGE/include/ege/music.h
	EGE/include/ege/sys_edit.h
	EGE/include/thread_queue.h
	EGE/src/ege/base.cpp
	EGE/src/ege/color.cpp
	EGE/src/ege/gapi.cpp
	EGE/src/ege/global.cpp
	EGE/src/ege/global.h
	EGE/src/ege/graphics.cpp
	EGE/src/ege/image.cpp
	EGE/src/ege/image.h
	EGE/src/ege/input.cpp
	EGE/src/ege/music.cpp
	EGE/src/ege/proc.cpp
	EGE/src/ege/random.cpp
	EGE/src/ege/text.cpp
	EGE/src/ege/time.cpp
	egeclock/egeclock.cpp
	egejulia/egejulia.cpp
	egenet/egenet.cpp
	egetriangle/egetriangle.cpp
	egetypegame/egetypegame.cpp
	graphicstest/maintest.cpp
	setup/setup.cpp
	test_setviewport/test_setviewport.cpp
	tutorial/tutorial.cpp

+++ EGE/include/ege/fps.h
@@ -21,11 +21,13 @@ public:
 		enable({});
 	}
 
-	void onDraw(IMAGE* pimg) const
+	void
+	onDraw(IMAGE* pimg) const
 	{
 		char str[16] = "fps        ", *pstr = str;
 		double fps = getfps() + 0.005;
-		int a = (int)fps, b = (int)((fps - a) * 100);
+		int a = fps, b = (fps - a) * 100;
+
 		while(*pstr)
 			++pstr;
 		pstr[-1] = (char)(b % 10 + '0');
